progress-persistence-handler: Ensure we release application
We wait for ::finished to release the application hold.

NautilusProgressInfo emits this signal in a idle callback. But if it
gets destroyed in the meantime, the idle source is removed. As a
result, the ::finish signal never gets emitted and we keep holding the
application forever.

So, keep it alive while we wait for the ::finished signal.