files-view: Drop .compare_files() vfunc
It gets called at a time when the new files haven't made it into the
view mode yet. So, we create a pair of dummy items for the sorter to
work on, and then throw them away.

Creating and destroying object instances at every step in the sorting
algorithm wastes a lot of cycles while loading large folders, with
visible performance impact.

So, instead of sorting before the view items are created, do it when
they are created, right before they are added to the model's internal
list store.