list-base: Don't modify selection model's internal bitset
In real_set_selection, we call gtk_selection_model_get_selection()
to be able to efficiently update the selection (we don't care about
unselected items that we aren't going to select).  That bitset,
per the documentation, shouldn't be modified.  This bug (which made
this function effective useless) went  unnoticed because we were
setting the selection elsewhere by poking the internals of the
listbase in set_focus_item.  We don't want to do that anymore, and
this bug became apparent.

Make a copy of the bitset, and rename the existing bitsets to clarify
what we are doing.  Also, set the selection after setting the
focus just in case that ever conflicts in the future.