files-view: Only set discrete scrolling when ctrl is held
The scroll controller on the scrolled window in files view
is set to discrete in order to accomodate for when CTRL+scroll
(zoom) is performed.  This however, causes problems with
the scrolls that are propogated and have delta's <1
(especially when they are  much much less 1).  So instead
let's set the discrete flag in scroll-begin only when CTRL
is held and remove it in scroll-end.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2511