error-reporting: Pass parent window to show_dialog()
For a very long time, we've been passing NULL as the
parent window to show_dialog() for failed rename
operations.  This is in a deeply nested callback, so
rather than trying to pass the window all the way back
just get the current active window, which very likely is
the correct window given how quickly this operation will
fail.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/1199