tracker-utilities: Make tracker connection async
Blocking on a sync connection to tracker on demand creates a noticeable
delay under normal circumstances and can be very problematic when
tracker is not responding.  Create the connection async, worst case
scenario the connection is not ready, and then we fall back to our other
search engine providers.  Unless tracker is broken anyways, that is
pretty unlikely.