list-base: Fix non-local DnD in x11
We are using a workaround to get the preferred action in x11
where we check the preferred action of the drag.  This works for local
drags, but obviously not non-local drags.  The workaround is necessary
because of a bug in gtk.  Let's just default back to COPY for non-local
drags.