list-base: Reimplement get_selection_for_file_transfer
The code for get_selection_for_file_transfer was abstracted from
view_icon_controller which didn't account for the issues of selecting
files with tree expanders.  Reimplement get_selection_for_file_transfer
to exclude children when the immediate parent is selected.  As noted in
7ac81c87c84149b421a6869837deb852876e3d12 this is not a perfect solution
because other ancestors (i.e. grandparents) could be selected, but at least
it matches the behavior we had previously.