window-slot: Handle archive URI as regular file if not the default handler
When nautilus is asked to handle a regular file (e.g. its URI being passed as command line argument),
we open its parent folder and select the file. But for archives, we want to extract them immediately 
instead (this is convenient when e.g. handling opening a downloaded file from the web browser).

However, we don't actually extract the archive unless we support are the default handler for that MIME type,
in which case we don't select it as a regular fiel either, which is inconsistent.

So, make sure we only handle archives specially if we actually meant to extract them.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2274