list-view: Reimplement Right/Left keys for tree expander
Tree expander uses SHIFT+Right/Left because the right/left
keys may be used to switch focus between columns.  We are just showing
labels, so there's no use for the Right / Left keys.  Also,
Nautilus has special behavior for the left key when the folder
is already collapsed (or a file which cannot be expanded) where it
selects the parent folder (if available). Bring back this behavior
with GtkTreeExpander and improve it for RTL languages.