window: Avoid crash when closing window
We remove the pad controller before calling gtk_window_destroy() to
break a dependency cycle.

However, if for some reason the controller has already been removed
by the time we reach that part of our code, we access a pointer to
and already destroyed object.

Try to avoid this with a weak pointer, such that we skip the step
if the pointer is already NULL.

Presumably fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2572