properties-window: Show FAT instead of msdos
We shouldn't be using filesystem::type, which isn't meant for UI, as
discussed in https://gitlab.gnome.org/GNOME/nautilus/-/issues/98

But a proper fix is going to take some time and possibly introduce a
new dependency (udisks2).

So, until we fix that issue, let's workaround the common outrageous
case of a FAT filesystem being captioned as "Msdos filesystem".

https://gitlab.gnome.org/GNOME/nautilus/-/issues/98