list-base: Move "cursor" when item removed
When an item is deleted, focus moves to the next item.

But if the previous item was not only focused by also selected, the
expected behavior is for the newly focused item to also become the
new selected item.

I believe this should be fixed in GtkListBase itself, but let's add
a workaround for now.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2489
... and provides a partial workaround for https://gitlab.gnome.org/GNOME/nautilus/-/issues/2489