file-operations: Always unlock worker thread on GtkDialog::respose
The worker thread is blocked while waiting for user response on the
main thread, through a GtkDialog. It gets unblocked upon response.

Except... we have an early return on our ::response hanlder, which
causes the worker thread never to be unblocked. This comes from
38a04943f9029ed4e8c7481b05502454a82a2811
The apparent intention was to presenve the behaviour from when we were
would calling gtk_dialog_run() in a loop. Maybe it worked under GTK3,
but it definitely doesn't under GTK4.

Fortunately, this should be dead code, because data->ignore_close_box
is always FALSE. Yet, under irregular situations, it may be reached,
as the following report suggests. So, it's best to remove this code
which is both wrong and unused.

Closes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2604