grid-view: Restore 2-dimensional navigation from sushi
We used to rely on the  "move-cursor" keybinding signal to forward
arrow key movements from sushi to our files view.

But GtkListBase-derived widgets do not provide a replacement for
that keybinding signal. While we have a good enough solution for
list view (which is one-dimensional, so we can just iterate the
list), it's a regression for grid view (up and down do not jump
to next line, because the grid's layout is unknown to us).

As a workaround, poke into GtkListBase's internal keybindings and
trigger their action programatically.

Resolves: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2556