search-popover: Search other types description
Currently we search only the mime-type string, because it was more
convenient for programming. But it's obviously inconvenient for users,
because the string shown in the list (i.e. the localized description)
doesn't match the strings being searched. This becomes worse when the
language is not English.

Use a more advanced expression to search both the mime type string as
well as its localized description.

This fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2659