autorun-software: Keep running until response
When moving away from gtk_dialog_run() [0], we forgot this small
program is not part of the main application. So, without anything
blocking the `main()` function, it's going to quit right away, without ever drawing the dialog.

Credit goes to Will Thompson for diagnosing the issue, and GTK 
test suit for the while loop solution I've taken from there.

Closes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2209

[0] d87f5be77bb4ee03e363cb560aaefb00cf56b275