window-slot: Handle broken default-folder-viewer
0 became the invalid id in 7c30c311e619a4d1c1d6b96f2095e5d63cfe2478

GSettings may give us an the invalid view id (e.g. if the stored value
doesn't match any string defined in the schemas). This may happen if
the backend is corrupted or the installed schemas are outdated, and it
leaves the application in a broken state.

So, handle this case, warning the user about the issue but falling
back to a valid view id.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2576