dnd: Don't allow dropping an archive on itself
In `nautilus_dnd_get_preferred_action()` we check if the dropped file
and target file are equal, but only after the check if we are dropping
onto an archive.  The `g_file_equal()` check if wrapped in a
`dropped != NULL` check which turns out to be unnecessary because
we return in the very beginning when `dropped == NULL`

Move around the drop target action checks to simplify and
prevent dropping anything on itself.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2737