list-base: Don't change location in a dnd when hovering over itself
It doesn't make sense to change locations onto the same folder you
are dragging from.  The easiest way to tell if we are dragging
onto the source drag is when the drag_item_action is set to 0.
This has the added benefit that drag_item_action is also set to
0 when hover over a file.  We have been calling
`nautilus_files_view_handle_hover()` unnecessarily on file destination
(which ultimately had no effect).

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2670