clipboard: Make Nautilus Clipboard more resilient
We are making the assumption that the clipboard data is always
correctly formatted.  Given that the "x-special/gnome-copied-files"
doesn't necessarily have a defined specification and that some
other software programs seem to be passing bad data, we need to be
more careful about the assumptions we make.

Let's add a few more checks:

1) The first line must begin with "copy" or "cut"
2) All subsequent lines must either be a valid URI or an empty string
(after stripping out '\n')
3) If there are any errors in the clipboard, return an error rather
than an empty clipboard

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2539