file-operations: Select files after a overwriting on move/copy
After pasting and selecting "Replace" in the file conflict dialog,
the new files are not selected (or focused).  This happens because
in nautilus-files-view, we are waiting for the "add-files" signal,
which never emits because the files aren't added, they are changed.
Set the overwritten files as FALSE in the debuting_files so that
we don't incorrectly wait for "add-files".

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2669