list-base: Allow to tab directly out of view
Now that the views contain real widgets, Tab and Shift+Tab cycle
through every item and only allow leaving the view when at the
end or start of the list, respectively.

This is frustrating for directories with more than a handful of files.
Also, users expect to use arrow keys to cycle though items already.

So, allow to tab into and out of the view, but not inside.