file-operation: Apply FAT filename rules to samba shares
We check if a filesystem is FAT, NTFS, etc in order to apply windows
based filename rules, but if we are copying to a smb  share, those filesystem
checks aren't performed.  The simple solution is to check the "cifs"
filetype also, but this fails because the filesystem_type is never
checked since the smb share doesn't return the error INVALID_FILENAME.
Instead it returns INVALID_ARGUMENT (and NOT_DIRECTORY for '\').
Since we are not handling those errors anyways, to be safe let's handle
it the same way we do INVALID_FILENAME