column-chooser: Subclass AdwWindow instead of GtkBox
We were subclassing from GtkBox and then, list-view created a
list-view-column-editor which subclassed from AdwWindow to add
the column-chooser.  It's a lot of extra code / complexity for something
that is only used in list-view.  This also allows us to directly
use column-chooser within list-view.

Also drop the list description label, in preparation for the switch to GtkListBox.