window-slot: Fix restoring scroll position from history
Nautilus does not scroll correctly for forward or backward navigation.

This is because the behaviour of the action "list.scroll-to-item" does
not put the specified position at the top of the view, but rather
scrolls the minimum amount to bring it to view, which usually leaves it
at the bottom of the view.

So, instead save the last visible item to scroll to.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2720

(Ammended by António Fernandes <antoniof@gnome.org>)