grid-cell: Show tooltip for ellipsized names
The old canvas expanded filename over the next row when the item was
selected. This is impossible to replicate in GtkGridView due to the
next row getting drawn after (== on top of) the previous one.

This is perceived as a feature regression. So, let's aleviate it
using tooltips when the filename is ellipsized.


(cherry picked from commit beeb2da98f4213a1e225291ae6b6c379f85398d4)