column-chooser: replace arrow box with row menus
Instead of selecting a row and moving it with up/down buttons, add a
menu to each row, that allows moving the row up and down. This gives the
per-folder switch in the bottom bar more space and allows the rows to be
activatable.

The moving menu is implemented using actions. To keep track of which
row the action is called for, store the row  for which the menu is
opened via a GtkMenuButtonCreatePopupFunc callback.