name-cell: Don't open on hover expander icon.
GtkTreeExpander includes a built-in "expand on drag hover" feature.

But our own "open on drag hover" feature competes (and often wins)
over it, making it hard to reliably expand rows on hover without
opening the folder.

So, separate the content and the expander. Trigger "expand on drag
hover" only if over the expander, and "open on drag hover" only if
over the rest.