files-view: Replace REMOVE_FILE with REMOVE_FILES
Files are added in bulk to the view with ADD_FILES, but then removed
from the view with REMOVE_FILE, which takes a single file.  This is very
inefficient in the case where a large number of items are deleted. In
other words, if you delete a folder with 1,000 items, the process
completes very quickly, but if you delete the 1,000 items, Nautilus will
hang for an extended period of time. This can be a common problem if
the view is current showing Trash, and the the trash is emptied.