special-location-bar: Factor out Settings launch code
We are going to change from GtkInfoBar to AdwBanner.

So, the GtkInfoBar::response signal is going away. In preparation,
isolate the reusable parts of its callback in functions that can
be used as calbacks to the AdwBanner::button-clicked signal.

(Factored out by António Fernandes <antoniof@gnome.org> from Peter's
original commit.)