column-chooser: Allow settings custom or default columns
When a folder has custom columns (i.e. the file metadata is set),
then show a banner saying there are custom columns, with a button to
reset to default.  If there aren't custom columns, then there won't be
a banner, and instead there will be a switch to allow setting the
columns as custom for the folder instead of default.

Default columns should only be updated on "close" for the window to
prevent user confusion.