column-chooser: Move operations from listview to column-chooser
There's this complicated back and forth between the listview and
the column chooser.  Work happens in both and it gets confusing
and creates unnecessary complexity.  It also will cause problems
in a future commit where we may want to revert changes that were
made but the list-view doesn't know about that.

Have the column-chooser load in all of the information, and set
the information.  The list-view is merely notified of the changes
for display purposes.