list-base: Sort after file-changes
When renaming a file in list-view or grid-view, the renamed file keeps
its position (i.e. it isn't re-positioned into the new proper position)
until a refresh). There is actually a more generalized problem where
file changes on all attributes fail to cause the sort positions to be
updated. For example, sorting by date modified and touching a file
doesn't cause a re-sort.

The file-changes signal from nautilus-directory, simply gives us the
changed files for a given directory, without telling us what was changed
so there's not an efficient way of detecting this.  We do have a
::end-file-changes signal, that is only emitted once every batch of
changes, so let's call nautilus_view_model_sort() then.  If no sort
changes are required, this will finish quite fast, so it seems like
the best alternative.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2568