file-operations: Warn before copying/moving >4GiB file to FAT
Nautilus attempts to copy/move >4GiB files to a FAT volume and fail
once it reaches the 4GiB threshold.  The verify_destination() function
already scans all of the source files before a copy / move, so we can
add a 4GiB check with almost no overhead.  Add attributes to the
SourceInfo to track the largest file.  If that file exceeds 4GiB, warn
before initiating the operation.  Similar to the existing before of
copying to a volume with insufficient space, allow forcing the copy.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/1899