flatpak: Use -Dbattery_detection=none for tracker-miners
The flatpak jobs are broken currently, because the tracker-miners project
fails to build due to missing upower-glib depenedency. It is used to tweak
the behavior when running on battery. It used to be automatically disabled
until the following MR:
https://gitlab.gnome.org/GNOME/nautilus/-/merge_requests/1121. I suppose
it is safe to disable this functionality as our flatpak isn't meant for
production. Let's explicitely disable that to fix our pipeline.