window: fix incorrect string when moving files to trash
When moving multiple files to trash, the toast currently
says: "%d files deleted". However, "deleted" implies permanent
loss of data. But, in this case, we are just moving files to the trash.
This commit corrects the string to show the actual action:
"%d files moved to trash".

Also correct Translators comments to make it consistent with the action.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2830