general: Change libadwaita include path
The libadwaita pkg-config adds `-I/usr/include/libadwaita-1`
(or similar) to the compiler flags, which means that

  #include <adwaita.h>

is sufficient, and the currently used

  #include <libadwaita-1/adwaita.h>

is not necessary.

Not only is it not necessary, but it prevents nautilus from being
built with an uninstalled libadwaita build (either meson subproject
or otherwise). To fix that, remove the `libadwaita-1/` prefix.

The libadwaita test programs also only `#include <adwaita.h>`, and
according to Debian Code Search, that is the more popular version.

https://codesearch.debian.net/search?q=%23include+%3Cadwaita.h%3E&literal=1
https://codesearch.debian.net/search?q=%23include+%3Clibadwaita-1%2Fadwaita.h%3E&literal=1