file-operations: Exchange error messages used when deleting
There are two error messages that may be shown when the delete operation
fails. One message is generic and the another one mentions insuffient
permissions. However, the generic one is now shown in the case of the
`G_IO_ERROR_PERMISSION_DENIED` error. Let's exchange the messages to
use the second one in this case.