flatpak: Fix Tracker3 access to portal
At some point, the flatpak command line interface changed so it is not
possible to add multiple values to a same key with a single --add-policy
call. Doing this now results in a single value added, with the ';'
escaped, this breaks Tracker portal access and results in Nautilus resorting
to the built-in miner instance.

It is however possible to call --add-policy multiple times to make
a multi-valued key, this restores the intended behavior and provides
access to the two tracker:FileSystem and tracker:Documents graphs.
Adapt the flatpak manifest to do this.

Closes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2838