search-engine: Create queries as TrackerSparqlStatements
These are not reused nor proper yet, this will happen in following
commits. At the moment, we go first through the step of creating
a TrackerSparqlStatement, and execute the query from there.

In order to create queries matching the enabled filter features,
we now use a flag set to express these capabilities. These capability
flags will be used further in later commits in order to cache
statements and assigning statement parameters.

A (so far) peculiarity of these flags is that we coalesce filename
and content search within the SEARCH_FEATURE_TERMS flag, but only
enable the latter with SEARCH_FEATURE_CONTENT. This will gain
relevance in later commits moving to fts:match for both sides of
the data. So far the distinct fn:contains and fts:match paths receive
the same search terms (if any was provided).