pathbar: do not use finalized pathbar to bind menu model to popover
Nautilus fails critical assertions if the user opens then closes a
new window in rapid succession in a large folder.

This is because the GSource created by
nautilus_path_bar_set_templates_menu to bind menu model to popover is
not removed even if the pathbar has already been finalized by the user
closing the window.

Keep track of the source id and remove source if the pathbar is
being disposed of.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2867