query-editor: Do not artificially delay search
With the changes done in the previous commits, the Tracker search
engine can take quite a pounding without breaking a sweat. Worrying
about spamming search engines is no longer a concern with queries
done the right way, so we do not need to artificially delay search.

The goal is that search feels responsive even to a fast typist.

Since this delay also applied to showing the search entry after
starting type-to-search, this also fixes the apparent stall when
one started typing.