desktop: Add application/zstd to supported mime-types
Currently, it isn't possible to use Nautilus for the extraction of the
".zstd" and ".tar.zst" archives. However, it seems that this compression
is supported by the libarchive library nowadays. Let's add the
`application/zstd` and `application/x-zstd-compressed-tar` mime-types
to the list of supported. This will allow Nautilus to show the "Extract"
and "Extract to..."  actions. Also, bump the gnome-autoar dependency to
the version with a corresponding change.

Related: https://gitlab.gnome.org/GNOME/gnome-autoar/-/merge_requests/38
Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2860