window: Use AdwOverlaySplitView instead of AdwFlap
AdwFlap has the same issues as all of our other adaptive widgetry.
In order to get rid of the issues caused by it, we need new widgetry
that can be manually used with AdwBreakpoint.

This commit ports the main window to the new AdwOverlaySplitView
widget. This widget is a simplified alternative to AdwFlap, and
for our use case it's a drop-in replacement.