properties-window: Use return value from g_string_free
Currently, the `ignoring return value of ‘g_string_free_and_steal’
declared with attribute ‘warn_unused_result’` warning is shown during
build. This is because the `str` struct member is used instead of a
return value from the `g_string_free` function. Let's update the code
to use the return value in order to avoid this warning.