search-hit: Remove redundant code
The `Deprecated pre-processor symbol: replace with "G_ADD_PRIVATE"`
warning is shown during build currently. This is caused by the usage
of the `G_TYPE_INSTANCE_GET_PRIVATE` macro. However, the statement with
this macro doesn't have any effect. This looks to be an oversight in
the commit 335eabec. Let's drop the whole statement to get rid of this
warning.