list-view: Get proper backing-uri for tree expanders
When using tree expanders, we have special cases for the backing uri.
In GridView or ColumnView (without expanders), the backing uri is
always the current view directory.  With tree expanders, the backing
uri is either the parent of the currently selected item, or the
currently selected item (if it is an expanded directory).  The same
rules apply for multiple selection if there is a common parent.

This was the case in 42, and the code was dropped in
6708861ed174e2b2423df0500df9987cdaf2adc0, and as an oversight wasn't added
back in when tree expanders were reintroduced.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2888