sidebar: Prevent auto folder opening on hover when re-ordering bookmarks
In the sidebar, when dragging to re-order bookmarks,
other bookmarked folders are automatically opening while hovering
over them as it happens during a drag-and-drop operation.

This action should not be permitted because it is inconceivable to
drag-and-drop a bookmark into another bookmark.

Implement a check to prevent automatically navigating
into the bookmarked folders if they are hovered on while dragging
a bookmark with the intent to just re-order them.
It achieves this behavior by not adding a navigation hover timeout
if we are dragging a row (bookmark) in the sidebar.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2820