view: Use tab-behavior API
When we are in the view, we don't want to tab through all of the
items, we use arrow keys for that.  Tabbing through the list, allows
easier accessibility to reach other controls.  Use GTK_LIST_TAB_ITEM
so that we can tab between items in a column view. For examples, the
quote text in a search, or the star (in a future commit).