star-cell: Make star focusable in list view
Now that we are using tab-behavior, we'll be able to tab to the star
to star items in an accessible way.  The current star is a GtkImage
which doesn't deal well with what are are trying to do (it doesn't
even show the focus ring).  Switch to a GtkButton which automatically
will handle click and activate.