search-engine-tracker: Don't try to parse NULL time strings
TrackerCursor returns NULL if it doesn't have information on the date.
https://gnome.pages.gitlab.gnome.org/tracker/docs/developer/method.SparqlCursor.get_string.html

We try to parse NULL as a date string. This is obviously wrong and
gets us a precondition failure warning, as it should.

So, if a string is NULL, just skip it.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2160