files-view: Don't add non-existent files
Under certain circumstances (i.e. moving a file), we end up adding
a non-existent file. This causes selection problems because the
non-existent file is selected and then removed (causing another file
to be selected).  I'm not sure where the root cause is (between
directory-async, directory notify, etc), but lets simply not add
a file that we know shouldn't be added.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2828