window: Dismiss previous undo toast when undo action has changed
We present a new toast every time the undo action changed which
is confusing because you might have several undo toasts appearing
over each other.  To complicate matters, the previous undo toasts
lie (undo'ing would actually undo a different action).

There was a proposal in #2350 to batch undo toasts, but that's problematic
because our undo manager is designed to undo exactly 1 action,
we do not have an undo stack.

The more straightforward solution seems to be simply dismissing the
old undo toasts, when we popup a new undo toast.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2350