style: Drop view item color overrides
We've added them because the colors assigned by the default stylesheet
were not adequate as per design review: d7b03656210d5feb300d5df74db9cd4601306991

Libadwaita 1.3 updated the stylesheet to use the design-choosen colors
by default in views. [0]

So, we should drop our overrides which are now unnecessary and even
harmful (they may be out of sync). Also, require the stable release of
the dependency, because 1.3.alpha didn't include [0].

[0] https://gitlab.gnome.org/GNOME/libadwaita/-/commit/fd83834f837baeb2650c75630de10ea5cdf032f2