window: Get actual preferred action for AdwTabBar drop
We are using a workaround to get the preferred action on the AdwTabBar
extra drop, by setting data on the page within ::extra-drag-value.
This hack doesn't account for the changes in preferred actions using
the CTRL/SHIFT modifiers.  Use the new adw api which gets the
action directly from the GdkDrop.