file: Re-use local time zone when generating date string
Whenever g_time_zone_new_local() is called it has to check the file
system for changes, because the system's time zone might have changed
since the last call. This function was called multiple times for each
file when generating the date string, which was causing some unnecessary
overhead.

This changes the code to re-use the local time zone from the first call.