build: Bump gtk depencency micro version
It includes the GTK bugfix mentioned in 2e10ed11d960cb6343719d680e29d742dc86d664

As we have removed the workaround, we cannot accept a GTK
version without that bugfix, because that would cause a
regression.

And we can't restore the workaround either, because it
misbehaves when using the fixed GTK version.

This leaves us with no other choice than requiring the
fixed GTK version.