general: Use GtkSwitch active property consistently
app-chooser sets the state instead of active property.  Similar to
d7af60161d30c885ebab69c58b346896f1565387, we really meant to set active
(whether the switch is toggled) and not the state (i.e. the color of the
switch).  Use gtk_switch_set_active()

In properties we've already fixed this, but there's a lingering call
to gtk_switch_get_state(). This is probably fine because state is
in sync with active when the default handler is used, but let's get
:active, to be consistent and safe.