list-view: Apply selection click workaround only if gtk < 4.10
In Nautilus 44.1 we removed the workaround completely because the
underlying problem was fixed in gtk 4.10.2.  We left the workaround
in 43.4 because the gtk patch wasn't cherrypicked in 4.8 (and
Nautilus 43 generally uses 4.8).  As it turns out, sometimes Nautilus
43 might have a patch gtk version >= 4.10.2 (i.e. on a rolling release
distro).  Add a runtime check to determine whether the workaround
should be applied.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2940