query-editor: Don't trigger search once cleared
We trigger search after a short time interval after the user stops
typing, as a performance saving measure.

To implement that behaviour, we need to ignore text changes that
are done programatically and not by the user typing. In that case,
we don't setup the timer that triggers the search.

But in that case we also don't clear a previously set timer. Although
the previous commit has already fixed the visible effects of this bug,
the timed callback is still called, which may cause other unknown bugs.

So, let's always clear the timer when the search text changes, even
if programatically, such as when clearing the text when closing search.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2875